\name{ozone}
\alias{ozone}
\title{
Ozone concentration, downtown L.A., 1955.1-1972.12
}
\description{
Ozone concentration, downtown L.A., 1955.1-1972.12
in parts per hundred million
}
\usage{data(ozone)}
\format{A ts object containing monthly data for 1955.1-1972.12.}
\source{
Intervention Analysis with Applications to Economic and Environmental Problems, 
G. E. P. Box and G. C. Tiao,
Journal of the American Statistical Association,
Vol. 70, No. 349 (Mar., 1975), pp. 70-79.
}
\keyword{datasets}
